requirements.txt
flask
gunicorn
Flask-Cors==3.0.10
pymongo
python-dotenv
requests